Game Finished. Needs saving and cloud

+++ Assets/Scripts/Models/Crop.cs
@@ -8,9 +8,9 @@ namespace Assets.Scripts.Models
 {
     class Crop
     {
-        public int GrowTime { get; set; }
-        public int Age { get; set; }
-        public int Value { get; set; }
+        public float GrowTime { get; set; }
+        public float Age { get; set; }
+        public float Value { get; set; }
         public Material Material { get; set; }
         public Mesh Mesh { get; set; }
     }
